+++ cmd/kapacitord/run/server.go
@@ -10,7 +10,6 @@ import (
 	"runtime"
 	"runtime/pprof"
 	"strings"
-	"time"
 
 	"github.com/influxdata/influxdb/influxql"
 	"github.com/influxdata/influxdb/services/collectd"
@@ -62,7 +61,8 @@ type Server struct {
 
 	err chan error
 
-	TaskMaster *kapacitor.TaskMaster
+	TaskMaster       *kapacitor.TaskMaster
+	TaskMasterLookup *kapacitor.TaskMasterLookup
 
 	LogService      logging.Interface
 	HTTPDService    *httpd.Service
@@ -71,7 +71,7 @@ type Server struct {
 	ReplayService   *replay.Service
 	InfluxDBService *influxdb.Service
 
-	MetaClient    *metaclient
+	MetaClient    *kapacitor.NoopMetaClient
 	QueryExecutor *queryexecutor
 
 	Services []Service
@@ -99,7 +99,7 @@ func NewServer(c *Config, buildInfo *BuildInfo, logService logging.Interface) (*
 		hostname:      c.Hostname,
 		err:           make(chan error),
 		LogService:    logService,
-		MetaClient:    &metaclient{},
+		MetaClient:    &kapacitor.NoopMetaClient{},
 		QueryExecutor: &queryexecutor{},
 		Logger:        l,
 	}
@@ -119,7 +119,9 @@ func NewServer(c *Config, buildInfo *BuildInfo, logService logging.Interface) (*
 	s.Logger.Printf("I! ClusterID: %s ServerID: %s", s.ClusterID, s.ServerID)
 
 	// Start Task Master
-	s.TaskMaster = kapacitor.NewTaskMaster("main", logService)
+	s.TaskMasterLookup = kapacitor.NewTaskMasterLookup()
+	s.TaskMaster = kapacitor.NewTaskMaster(kapacitor.MainTaskMaster, logService)
+	s.TaskMasterLookup.Set(s.TaskMaster)
 	if err := s.TaskMaster.Open(); err != nil {
 		return nil, err
 	}
@@ -206,7 +208,7 @@ func (s *Server) appendInfluxDBService(c []influxdb.Config, defaultInfluxDB, htt
 
 func (s *Server) initHTTPDService(c httpd.Config) {
 	l := s.LogService.NewLogger("[httpd] ", log.LstdFlags)
-	srv := httpd.NewService(c, l)
+	srv := httpd.NewService(c, l, s.LogService)
 
 	srv.Handler.MetaClient = s.MetaClient
 	srv.Handler.PointsWriter = s.TaskMaster
@@ -225,7 +227,7 @@ func (s *Server) appendTaskStoreService(c task_store.Config) {
 	srv := task_store.NewService(c, l)
 	srv.StorageService = s.StorageService
 	srv.HTTPDService = s.HTTPDService
-	srv.TaskMaster = s.TaskMaster
+	srv.TaskMasterLookup = s.TaskMasterLookup
 
 	s.TaskStore = srv
 	s.TaskMaster.TaskStore = srv
@@ -240,6 +242,7 @@ func (s *Server) appendReplayService(c replay.Config) {
 	srv.HTTPDService = s.HTTPDService
 	srv.InfluxDBService = s.InfluxDBService
 	srv.TaskMaster = s.TaskMaster
+	srv.TaskMasterLookup = s.TaskMasterLookup
 
 	s.ReplayService = srv
 	s.Services = append(s.Services, srv)
@@ -582,26 +585,6 @@ type tcpaddr struct{ host string }
 
 func (a *tcpaddr) Network() string { return "tcp" }
 func (a *tcpaddr) String() string  { return a.host }
-
-type metaclient struct{}
-
-func (m *metaclient) WaitForLeader(d time.Duration) error {
-	return nil
-}
-func (m *metaclient) CreateDatabase(name string) (*meta.DatabaseInfo, error) {
-	return nil, nil
-}
-func (m *metaclient) Database(name string) (*meta.DatabaseInfo, error) {
-	return &meta.DatabaseInfo{
-		Name: name,
-	}, nil
-}
-func (m *metaclient) Authenticate(username, password string) (ui *meta.UserInfo, err error) {
-	return nil, errors.New("not authenticated")
-}
-func (m *metaclient) Users() ([]meta.UserInfo, error) {
-	return nil, errors.New("no user")
-}
 
 type queryexecutor struct{}