+++ integrations/helpers_test.go
@@ -3,6 +3,7 @@ package integrations
 import (
 	"errors"
 	"fmt"
+	"io"
 	"log"
 	"net/http"
 	"net/http/httptest"
@@ -12,7 +13,6 @@ import (
 
 	client "github.com/influxdata/influxdb/client/v2"
 	"github.com/influxdata/influxdb/influxql"
-	"github.com/influxdata/influxdb/services/meta"
 	"github.com/influxdata/kapacitor"
 	"github.com/influxdata/kapacitor/udf"
 	"github.com/influxdata/wlog"
@@ -128,6 +128,17 @@ type LogService struct{}
 func (l *LogService) NewLogger(prefix string, flag int) *log.Logger {
 	return wlog.New(os.Stderr, prefix, flag)
 }
+func (l *LogService) NewRawLogger(prefix string, flag int) *log.Logger {
+	return log.New(os.Stderr, prefix, flag)
+}
+
+func (l *LogService) NewStaticLevelLogger(prefix string, flag int, level wlog.Level) *log.Logger {
+	return log.New(wlog.NewStaticLevelWriter(os.Stderr, level), prefix, flag)
+}
+
+func (l *LogService) NewStaticLevelWriter(level wlog.Level) io.Writer {
+	return wlog.NewStaticLevelWriter(os.Stderr, level)
+}
 
 type UDFService struct {
 	ListFunc   func() []string
@@ -168,27 +179,3 @@ func (d deadman) Threshold() float64      { return d.threshold }
 func (d deadman) Id() string              { return d.id }
 func (d deadman) Message() string         { return d.message }
 func (d deadman) Global() bool            { return d.global }
-
-type metaclient struct{}
-
-func (m *metaclient) WaitForLeader(d time.Duration) error {
-	return nil
-}
-
-func (m *metaclient) CreateDatabase(name string) (*meta.DatabaseInfo, error) {
-	return nil, nil
-}
-
-func (m *metaclient) Database(name string) (*meta.DatabaseInfo, error) {
-	return &meta.DatabaseInfo{
-		Name: name,
-	}, nil
-}
-
-func (m *metaclient) Authenticate(username, password string) (ui *meta.UserInfo, err error) {
-	return nil, errors.New("not authenticated")
-}
-
-func (m *metaclient) Users() ([]meta.UserInfo, error) {
-	return nil, errors.New("no user")
-}